+++ plugin/nube/protocals/edge28/config.go
@@ -1,12 +1,12 @@
 package main
 
-type Config struct {
-	EnablePolling bool `yaml:"enable_polling"`
-}
+import "github.com/NubeIO/flow-framework/plugin/nube/protocals/edge28/config"
 
 func (inst *Instance) DefaultConfig() interface{} {
-	return &Config{
+	return &config.Config{
 		EnablePolling: true,
+		LogLevel:      "NONE", // DEBUG or ERROR
+		PollRate:      5,      // in seconds
 	}
 }
 
@@ -14,8 +14,8 @@ func (inst *Instance) GetConfig() interface{} {
 	return inst.config
 }
 
-func (inst *Instance) ValidateAndSetConfig(config interface{}) error {
-	newConfig := config.(*Config)
+func (inst *Instance) ValidateAndSetConfig(c interface{}) error {
+	newConfig := c.(*config.Config)
 	inst.config = newConfig
 	return nil
 }