+++ plugin/nube/protocals/rubixio/jsonschema/device.go
@@ -1,0 +1,18 @@
+package jsonschema
+
+import "github.com/NubeIO/lib-schema/schema"
+
+type DeviceSchema struct {
+	UUID        schema.UUID        `json:"uuid"`
+	Name        schema.Name        `json:"name"`
+	Description schema.Description `json:"description"`
+	Enable      schema.Enable      `json:"enable"`
+	Host        schema.Host        `json:"host"`
+	Port        schema.Port        `json:"port"`
+}
+
+func GetDeviceSchema() *DeviceSchema {
+	m := &DeviceSchema{}
+	schema.Set(m)
+	return m
+}